+++ src/components/post/post-more-menu.jsx
@@ -15,6 +15,7 @@ import styles from '../dropdown-menu.module.scss';
 import { format } from '../../utils/date-format';
 import { MenuItemIconic as Iconic } from './menu-item-iconic';
 import { MenuItemTranslate } from './menu-item-translate';
+import { MenuItemNotifyOfAllComments } from './menu-item-notify-on-all-comments';
 
 // eslint-disable-next-line complexity
 export const PostMoreMenu = forwardRef(function PostMoreMenu(
@@ -120,7 +121,14 @@ export const PostMoreMenu = forwardRef(function PostMoreMenu(
         </ButtonLink>
       </div>
     )),
-    [<MenuItemTranslate key="translate" type="post" id={postId} doAndClose={doAndClose} />],
+    [
+      <MenuItemNotifyOfAllComments
+        key="notify-of-all-comments"
+        postId={postId}
+        doAndClose={doAndClose}
+      />,
+      <MenuItemTranslate key="translate" type="post" id={postId} doAndClose={doAndClose} />,
+    ],
     [
       isDirect && !isOwnPost && (
         <div className={styles.item} key="leave-direct">