+++ src/components/user.jsx
@@ -1,5 +1,5 @@
 /* global CONFIG */
-import { useEffect, useMemo } from 'react';
+import { useCallback, useEffect, useMemo, useState } from 'react';
 import { Helmet } from 'react-helmet';
 import { connect } from 'react-redux';
 import * as _ from 'lodash-es';
@@ -19,6 +19,7 @@ import Breadcrumbs from './breadcrumbs';
 import ErrorBoundary from './error-boundary';
 import UserProfile from './user-profile';
 import UserFeed from './user-feed';
+import { ButtonLink } from './button-link';
 
 const UserHandler = (props) => {
   // Redirect to canonical username in URI (/uSErNAme/likes?offset=30 → /username/likes?offset=30)
@@ -44,6 +45,18 @@ const UserHandler = (props) => {
     props.viewUser.isLoading,
     props.viewUser.username,
   ]);
+
+  const [forceShowContent, setForceShowContent] = useState(false);
+  const displayPosts = useCallback(() => setForceShowContent(true), []);
+
+  const allowToPost =
+    !CONFIG.privacyControlGroups.hidePosts ||
+    !CONFIG.privacyControlGroups.groups[props.viewUser.username];
+  const showContent =
+    forceShowContent ||
+    !CONFIG.privacyControlGroups.hidePosts ||
+    !CONFIG.privacyControlGroups.groups[props.viewUser.username];
+  const controlledPrivacy = CONFIG.privacyControlGroups.groups[props.viewUser.username]?.privacy;
 
   const nameForTitle = useMemo(
     () =>
@@ -87,6 +100,7 @@ const UserHandler = (props) => {
           <UserProfile
             {...props.viewUser}
             {...props.userActions}
+            canIPostHere={props.canIPostHere && allowToPost}
             user={props.user}
             sendTo={props.sendTo}
             createPost={props.createPost}
@@ -98,7 +112,25 @@ const UserHandler = (props) => {
           />
         </div>
 
-        <UserFeed {...props} />
+        {showContent ? (
+          <UserFeed {...props} />
+        ) : (
+          <div className="alert alert-warning">
+            <p>
+              This is a <strong>{props.viewUser.username}</strong> group page. This is a special
+              group that is used for changing the visibility of posts in other feeds to “
+              {controlledPrivacy}”.
+            </p>
+            <p>
+              The posts in this group itself are random and are not organized around any specific
+              topic.
+            </p>
+            <p>
+              Do you still want to see them?{' '}
+              <ButtonLink onClick={displayPosts}>Yes, show me posts</ButtonLink>
+            </p>
+          </div>
+        )}
       </ErrorBoundary>
     </div>
   );